+++ ControlDeMetas.DAL/Contracts/IRepository.cs
@@ -10,7 +10,7 @@ namespace ControlDeMetas.DAL.Contracts
     public interface IRepository<TEntity> where TEntity : class
 {
         Task<IEnumerable<TEntity>> GetAllAsync();
-        Task<TEntity> GetByIdAsync(int id);
+        Task<TEntity> GetByIdAsync(long id);
         Task<TEntity> AddAsync(TEntity entity);
         Task<TEntity> UpdateAsync(TEntity entity);
         Task<bool> DeleteAsync(int id);
